pyproject.toml
[tool.poetry]
name = "server-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.10,<=3.12"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
fastapi = { version = "^0.110.0" }
uvicorn = { extras = ["standard"], version = "^0.29.0" }
alembic = "^1.13.1"
sqlalchemy = "^2.0.29"
python-dotenv = "^1.0.1"
click = "^8.1.7"


[tool.poetry.group.dev.dependencies]
httpx = "^0.27.0"
pytest = "^8.1.1"
pre-commit = "^3.7.0"
starlette = "^0.37.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# if you want to use a mirror, uncomment the following lines
[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
# priority = "default"

[tool.isort]
line_length = 88
